Fix pbench-trafficgen to record metadata properly
Fixes #1079.

Up until this point all pbench-trafficgen results are not indexable
because the metadata.log file is not properly setup. We modify the
pbench-trafficgen script to behave like pbench-uperf does.

This also meant adding support for --sysinfo, but using a default of
"none" because that was the current behavior of pbench-trafficgen.
Now users who want/need to collect system information can add the
--sysinfo parameter as well.

Lastly, while reviewing pbench-trafficgen we fixed the user of the
total_iterations variable so that it is reported properly in the
logs, fixed a bad reference to $benchmark_name, exported the proper
$benchmark environment variable as all the other bench scripts do,
and cleaned up the location of what happens when so that it flows
similarly to uperf.

Along the way, we also addressed some minor issues that were noticed
in uperf and moongen when comparing them against trafficgen.