Consistently handle bench scripts options & help
This is an attempt to make sure that all active and supported pbench
bench scripts have common handling for:

  * the --help option
  * bad options
  * bad option values for --tool-group and --sysinfo

To make this more consistent we added a single function in agent/base to
check both --sysinfo and --tool-group parameters uniformly.

We also reduce the number of unit tests writing one wrapper script which
invokes all the benchmarks with the combination of parameters.

Along the way, we made sure pbench-linpack actually works while fixing
other minor problems and inconsistencies.