Ensure indexing properly cleans up extracted tar balls
Prior to this change, if the indexing job has more than one tar ball to
index, index-pbench will extract each tar ball it works on, leaving the
extracted contents to be cleaned up by pbench-index. This works well for
one or two tar balls, but if we have a large set of tar balls to work
through this can be problematic.

This change makes it so that index-pbench properly cleans up the
extracted tar ball before exiting.  We modify pbench-index to give a
dedicated sub-directory to index-pbench so that we can detect problems.
We also modify the unit test harness to detect when any server script
leaves data around in its given TMPDIR.