Rearrange sar-postprocess code to match order in file
In order to prepare for upgrading to a new version of sysstat, we have
rearranged the order of the code to match the order the sections are emitted
by the "sar -A".  This makes it easier to find where a section is processed.

We also only compare full line regular expressions only when in a given
mode, and add next statements to avoid further comparisons that are not
necessary.