Restructure selected dashboard components in redux store
For handling persistence of the redux store, selected configuration
data needs to be stored in the global reducer while the dashboard
namespace is blacklisted to avoid local caching.

Previously, caching of the dashboard namespace was causing conflicts
with an edge case where server side data was being updated but was not
reflected in persisted user sessions.