Eliminate some variables from pbench-server-default.cfg
Fixes #1195

Eliminate the following variables from pbench-server-default.cfg:

- pbench-incoming-dir
- pbench-results-dir
- pbench-users-dir
- pbench-static-dir

All of these are now fixed relative to pbench-top-dir:

- pbench-incoming-dir = <pbench-top-dir>/public_html/incoming
- etc.

So the only "free" variable left is the top-level directory.

This requires minor changes to a couple of the activate scripts
and the PbenchConfig class.

The primary reason for this is to simplify the Ansible roles
for installing the pbench server. This PR will be a dependency
for that.

Simplify the handling of the shim reception areas.

We now define a prefix for the directory in the config file.  The
prefix does not include the version, which is tacked on specifically
when needed. This required modifications to the shim scripts and a
couple of the activate scripts.

Split host_info_url into host and path pieces

pbench-move-results retrieves the whole URL, but
pbench-server-activate-setup-results-host-info only needs the path, so
we make that explicitly available as host-info-path-prefix (without a
version number), so that it does not have to parse the URL to retrieve
the path piece.