Eliminate some variables from pbench-server-default.cfg
Fixes #1195

Eliminate the following variables from pbench-server-default.cfg:

- pbench-incoming-dir
- pbench-results-dir
- pbench-users-dir
- pbench-static-dir

All of these are now fixed relative to pbench-top-dir:

- pbench-incoming-dir = <pbench-top-dir>/public_html/incoming
- etc.

So the only "free" variable left is the top-level directory.

This requires minor changes to a couple of the activate scripts
and the PbenchConfig class.

The primary reason for this is to simplify the Ansible roles
for installing the pbench server. This PR will be a dependency
for that.