Server installation ansible playbooks and roles.
Fixes #1218

These playbooks implement a complete install and setup of one or more
pbench servers.

In addition to installing the pbench-server RPM and the config file
for each server, they do all the work that the pbench-server-activate-*
scripts do.

Most of the scripts haven been "ansiblized": they use appropriate
ansible modules in order to carry out their operations.  The one
exception is the pbench-server-activate-create-crontab role: the
ansible script just executes the command remotely.  The reason for
that is that reimplementing requires fetching too many variables from
the config file in the server, complicating and slowing down the
playbook.  The other scripts already do a fair amount of such fetching
of configuration settings(see the pbench-server-vars role) but the
crontab role would double the number of variables.

This commit makes all the server/bin/pbench-server-activate-* scripts
(*except* for pbench-server-activate-create-crontab) obsolete. They
will be removed in a subsequent PR, once we are sure that the ansible
playbooks are reliable.

The pbench-server-activate-setup-web-server role uses the synchronize
module, a wrapper around rsync, so the role ensures that rsync is
installed.

Since we are synchronizing a directory on each remote host, each
remote host has to be able to login passwordlessly to itself.

A README.org file has been added, providing complete documentation
of the basic features (and some more advanced features) of the
playbooks.

Changes requested on review (mostly):

- Change the install state of the RPM from "present" to "latest".

- Use pbench_install_dir variable everywhere.

- Fix dependencies.

- Fix README.org.

- Define all variables except configurl, into role defaults.  The
  pbench_configuration_environment (and its command-line alias, cenv)
  has been eliminated.

- Define a new role, pbench-server-config, that does the server
  configuration after the config file is installed. This new role
  basically calls the old roles, which still exist, as subroutines. As
  a result, the pbench-server-install playbook has been simplified.

- The template for the pbench test repo now uses a variable to set
  whether the repo will be enabled. By default, the variable is set to
  0 (disabled), but it is now possible to set it to 1 from the command
  line, which is convenient for testing.

- The pbench-config-install role has been renamed to
  pbench-server-config-install: it was meant to be identical to the
  corresponding pbench-agent playbook, but they have diverged.