Reduce server unit tests time
The server unit tests 7.13 and 7.14 were indexing 1 million+ proc-int
documents.  This cuts those counts down to about 24,000+, reducing the
time considerably.

Further, for production, we add logging for the tool data indexing to
know when it starts and stops so that we can debug further any bad
indexing performance behaviors.

To add the logging, we refactored the ToolData class to make sure the
path for the tools is calculated consistently in one place so that it
can be used in the log output.