Refactor agent to use envi for perl module paths
We had too much logic to support the unit tests in all the various perl
modules making it complicated to maintain. We strip all that out,
replacing those mechanisms with just environment variables provided by
the `base` script.

This simplifies much of the code and gives us a way to add other perl
module paths in the future.

We define the new `pbench_lib_dir` environment variable that the CDM
code already uses, adding it to the `base` script, and then we take
advantage of that by moving the common perl modules that were in
`tool-scripts/postprocess` to `agent/lib`.

We also fix up the config files `pbench-agent.cfg.example` and
`pbench-agent-default.cfg` so that all of the defaults exist in the
default config file, and so that the unit tests can leverage the use of
the example config file for their operation. We then modified all
references to the `pbench_install_dir` in the config file to be
`install-dir` to be clear we weren't try to reference the environment
variable.

Other changes:

 * We fixed up the logging mechanisms to emit all the arguments given on
   the call, this caused a change to the `pbench-dbench` `test-02`
   output

 * Along the way, while debugging unit test failures as a result of
   these changes, we added some logging for error paths to
   `pbench-register-tool-set` and `pbench-register-tool`, as well as
   some minor code changes and clean up to make it easier to understand
   the logic

    *  We left a `FIXME` in the comments asking why we unconditionally
       register the `perf` tool even when it is not requested