Fix use of sosreport for version 3.5.1 and later
Starting with version 3.5.1, the `general` plugin was removed, having
it functionality moved into other existing plugins, and two new plugins
added, `date` and `host`.  The collection of `/proc/stat` was in the
`general` plugin, but was moved to the `process` plugin.  We don't add
the `process` plugin at this point because it was not on our original
list.  A separate commit should propose the addition of such a plugin.
In the mean time, we add the collection of the `date` and `host`
plugins now.

The `lsbrelease` plugin was renamed to `release`.

See also the sosreport commit: [971b958](https://github.com/sosreport/sos/commit/971b9581779da20384f0a4d8de5177c0b87d6892)