Add support for distributed shared filesystems w fio
We add documentation explaining how to use pbench-fio on a distributed,
shared, filesystem, providing an example job file. There is a limitation
for support of this feature as it has not been exhaustively tested with
all pbench-fio options.

We also update the fio documentation for RBD, Cephfs tests.

To enable support for distributed, shared filesystems, we unified the
handling of the `-c` and `--client-file` behaviors always using a
client file when invoking `fio`.

For distributed fs testing, we raise `max_jobs` by number of jobs per
host considering the targets specified.

If the user wants to specify multiple fio processes per client/pod with
a shared file system, they must uncomment the `numjobs=` line in the
new shared fio job file.