Refactoring of fio_create_jobfile
Use argument numbers in the same order as the caller, and add a comment to
the call sight to help ensure the order is correct.

Use `error_log` instead of `debug_log` to report failures related to creating
the fio job file.

Explicitly check that the path to a fio job file is created successfully.