Move start/stop tools closer to fio command
The location of where tools are started and stopped are now close to
where the actual fio command is initiated.  The start-tools and
pre-iteration-script we moved just above the fio command execution, and
the postprocessing-tools was moved immediately after the tools were
stopped.

We have removed the check for the existence of a fio job file at the
start of `fio_run_job` since the caller has already assured us the file
exists.

We have removed the re-definition of `benchmark_bin` as `bench_cmd`
since that only served to obscure its use in the code.

Finally we have removed the redundant `mkdir` call for the creation of
the results directory, adding an explicit check that it successfully
created a directory.