Remove redundancy in pbench-fio operations
We now generate the `$benchmark_run_dir/fio-client.file` once and then
use that one file in all the `fio --client-file=` invocations.

We now generate the `fio.job` file once per iteration, instead of for
each sample (this makes it easier to wade through unit test output to see
if the behaviors are correct).

We have simplified the handling of `--postprocess-only` to reduce the
number of times it is checked in various parts of the code.

Using `--install` and `--postprocess-only=y` is incompatible, so we now
issue a message, post the `usage`, and exit with an error.

We now post the `usage` message when exiting for `-c`|`--clients` and
`--client-file` incompatibilities.

We also obsolete the --remote-only option since it was unused.

We only create the `.iterations` file in the `$pbench_tmp` directory if
we are going to proceed with a run (even for `--postprocess-only=y`).