Reduce redundant test setup operations
This is a rework of the unit test infrastructure for bench-scripts to
reduce the amount of redundant operations performed for each test.

Now, all tests automatically have a `--config=test-##` parameter added
to their invocation, and the full pbench-agent.cfg file is used for them
as well.

We enhance the infrastructure to look for common files that tests might
leave as artifacts that should show up on the gold file output.

We add additional error checking for creating directories and files
during test setup.

We correct a bug with pbench-dbench so that it now properly emits the
iteration name.

We also remove trailing spaces from generated output, and add a missing
newline to the user-benchmark JSON file generated to make it look like
expected output in the gold files.