Rework pbench-agent ansible playbooks
Rework the ansible playbooks to make them as similar as possible
to the server ansible playbooks (PR#1199).

There are three playbooks: the main one is pbench-agent-install.yml
which installs the COPR repos, uses them to install the pbench-agent
RPM from COPR and then copies the config file and key file for the
environment requested through the cenv command line variable (if
this is not set, the default is "production").

Usage:

     ansible-playbook --user=root -i /path/to/inventory.file \
                      pbench-agent-install.yml \
                      [ --extra-vars='{cenv: <environment>}' ]

Two other playbooks use subsets of the defined roles:
pbench-repo-install.yml installs the COPR repos only. You can then
install the pbench-agent RPM manually.  In that case, the third
playbook can be used to install the config and key files afterwards.

An example inventory file, Inventory/pbench-agentf.hosts.example, is
provided. The second section of the file contains the definition of
various variables that are used by the playbooks.  These are set to
dummy values here, so a site admin should modify this file for a
particular installation and make it available to the users for that
site.

Users should then copy and mofify the example file, without touching
the variables section: they should just add their hosts to the first
section (possibly augmented with host varirable definitions,
e.g. ansible_python_interpreter, that might be need to work around
problems on some distros).