Add datalog unit tests; fix prometheus-metrics
We only use the infrastructure to test the prometheus-metrics-datalog
funtionality. Others can be added later as time and resources permit.

We also fix up the mock of time/sleep so that the unit tests will only
execute the datalog loop 3 times no matter what.

Finally, we fix `prometheus-metrics` to use `PYTHONUNBUFFERED=True`
so that we are less likely to lose data if killed while executing.