Add ability to test fio-postprocess-viz.py
We also fix up `fio-postprocess-viz.py` so that it accepts and
processes the same arguments as `fio-histo-log-pctiles.py` so that two
scripts will always behave the same way.

Further we add the output of the generated `results.html` files to the
unit test gold files in `test-30` so that we can be sure that
`fio-postprocess-viz.py` is generating correct HTML files.