Refactor to fix and clean up prom-met-pp
The previous way were converting prometheus metrics resulting in an
extra record when there were just `labels` in the original record. That
cause many extra records to be generated with bad data.

Second problem is that the `quantiles` and `buckets` arrays were not
processed sorted by the quantile or bucket value, which meant that it
was possible to get different unit test output ordering dependeing on
the version of python used.

In order to help debug the conversion process, we added to additional
fields to the generated records:

 * `_idx` - which is basically the "index" ID of the JSON document read
            from the source data file (aka, line number)

 * `_metric_idx` - which is the index ID of the metric rendered out
                   of a list of metrics for a given `_idx` original
                   JSON document (some JSON documents have more than
                   one metric in them)

Further, we rename the resulting field names:

 * `quantiles` : `quantile`
 * `buckets` : `bucket`

We also take this opportunity to ensure we handle errors in individual
JSON documents so that we process as many valid ones as possible.

Lastly, we unify the number conversion methods so that there is only
one method for integer and float number handling.