Fix bench-scripts/unittests to *not* create the benchmark run dir.
Bench-scripts are supposed to create their own benchmark_run_dir
directories since only they know what time they start, which is part
of the name of that directory.

The unittests script cheated and created the benchmark_run_dir
directory, since it knows the (fixed) time that we impose for
reproducible unit test results. This cheat actually hid bugs that were
introduced in the recent change to move the iterations file from the
tmp/ directory to the benchmark_run_dir directory (PR #1261).

This commit changes the unittests script and gets rid of the
benchmark_run_dir creation. That uncovered an additional bug in the
script: it was trying to cd to that directory unconditionally, in
order to dump its contents to the output file, but since the directory
no longer exists, the cd failed with an error message; the solution
was to make the cd conditional: if it succeeds then we dump the
contents, otherwise we do nothing.

The unittests script wanted to create the directory in the first
place, in order to initialize the metadata.log file contained
therein. This initialization is now done in the test-bin/mock-cmd
script (pbench-metadata-log is mocked, but it should probably not be:
that might be done in a later commit). That requires that mock-cmd
have access to the benchmark_run_dir variable: all bench-script export
that, except pbench-trafficgen, so it too was changed to export the
variable.

Finally, these changes caused some unit test output changes: the
directory and the metadata.log might not exist at all if the bench
script is run with erroneous options.

Note that not all the unit tests pass: the bug that was introduced by
PR #1261 is properly caught as a unit test failure. The next commit is
going to fix the problematic cases so that all the unit tests pass.