Fix server reports index, and report status
We have fixed the report_status() python method to index 50 MB chunks
of the given file instead of trying to send the entire file. When we
encounter cases of error files being larger than 100 MB, Elasticsearch
will reject it by responding with a 502 error but no reason given. So
we now only send at most 50 MB of the file at a time, breaking up into
as many chunks as necessary, adding the total size of the original
data, the number of chunks, and which chunk the current document is
as metadata.

We are now using an index name pattern for pbench reports that looks
like: `<index prefix>.server-reports.<YYYY>-<MM>`. This was simply a
change to use a dot (`.`) instead of a dash (`-`) between the prefix
and the `server-reports` string.

This change to the mapping and indexing strategy requires us to bump
the major indexing version.

We also remove the last hard-coded use of the "pbench-index" name.