Enable dispatch-states in the server default
Currently if a server is deployed without dispatch-states, everything
falls down.  Seems better to provide the base defaults instead of
requiring each server configuration to always provide a set of states
that really only the internal code knows about.

That is to say, this would enable us to consider removing this entirely
from the configuration files.