Remove move-unpacked; unpack additional states
First, we remove `pbench-move-unpacked`. It is no longer used in any of
our deployments, and it is a hassle to keep around the code. Removing it
now is less painful than trying to remove it later when other code
changes may be necessary.

Second, we now fix the order in which operations performed. The indexing
and the copying of sosreports now follows unpacking of tar balls. This
removes a race condition where `pbench-copy-sosreports` may miss seeing
sosreports because the tar ball is not unpacked. It also gives us the
opportunity to leverage the already unpacked tar balls for the indexing
step (planned for a future commit).