Remove user-metadata opt from server invocation
The pbench-index command has always been designed to index mulitple tar
balls from many different sources that have been pushed to the pbench
server.  The user-metadata option was added to the former index-pbench
python script which only indexed a given tar ball at a time, and so was
appropriate to that use case.  Removing it now avoids confusion later as
the code ages.