Various fixes to server side unit tests
The first small change is to invoke pbench-index with the long form of
its arguments to enhance the description of what the unit test does.

The second larger change is to fix the behavior of the unit tests so
that tests can exit before they are waited on, preventing errors flowing
from that condition from affecting the test outcome.