Add "@generated-by" indexing info for tracking
We add an `@generated-by` record with more context about who and what generated
a run document, and when it was generated. This will help us understand better
the code that generated a document when debugging problems.

We also bump the run document versions to cover this change, and to cover the
change previously made in PR #1295 for the path tokenizer addition.