Create a pbench "indexer" module
Rather than having a monolith `pbench-index.py` file, we move all the
logic for indexing a pbench result tar ball into a new module,
`lib/pbench/indexer.py`.  This gives us some flexibility going forward
for how to make changes to the indexing code, isolated from how the
indexing driver operates.

This should be a pure "lift-n-shift".  Some of the gold files changed
because the paths to the function names changed, but that should be it.