Break pbench-index into two indexers: run/result & tools
Since there can be a tremendous amount of tool data, we break the pbench
result indexing into two steps: run, table-of-contents, and result data
in one job, and tool data in another job.  This separation should allow
all the tar balls to be processed quickly to get their run data indexed
for the dashboard, allowing the tool data to follow later.