Enhanced status reporting for tracking
Now that we have a separate job for indexing "run", "table-of-contents",
and "result" data, and another job for indexing "tool" data, we can no
longer keep the "generated-by" tracking information in the "run"
document.  The tool data could be indexed at a much later time with a
different version of the software, or even on a different host.

We enhance the status reporting infrastructure, creating a new class,
`Report()`, which is constructed with all the meta data related to the
operation of a particular server-side script: host name, pid, gid, uid,
time it started running, version, git commit hash, etc.  The class has
a `post_status()` method to allow the caller to create a JSON document
that will be indexed into Elasticsearch recording the given status.

This new `post_status()` method returns the ID of that document so that
the caller can use its ID as the `@generated-by` field for all other
data indexed.  This allows us to put arbitrary information about the
process performing the indexing, without adding unneeded metadata to all
indexed documents, while allowing us to track down the origin of the
indexer.